requirements.txt
beautifulsoup4
python-decouple
requests
